requirements.txt
albumentations==0.4.5
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.1.2
streamlit==0.62.1